Bug #16934 Unsigned values > 2^63 (UInt64) cannot be used in prepared statements
Small correction to the previous commit.

+++ mysqlclient/parameter.cs
@@ -315,7 +315,7 @@ namespace MySql.Data.MySqlClient
                 case MySqlDbType.UInt32:
                     return (int)MySqlDbType.Int32 | 0x8000;
                 case MySqlDbType.UInt24:
-                    return (int)MySqlDbType.Int64 | 0x8000;
+                    return (int)MySqlDbType.Int32 | 0x8000;
                 case MySqlDbType.UInt16:
                     return (int)MySqlDbType.Int16 | 0x8000;
                 default:
